+++ src/EthernetWebServer.h
@@ -12,7 +12,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.8.3
+  Version: 1.8.4
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -25,17 +25,18 @@
   1.8.1   K Hoang      24/12/2021 Fix bug
   1.8.2   K Hoang      27/12/2021 Fix wrong http status header bug
   1.8.3   K Hoang      28/12/2021 Fix authenticate issue caused by libb64
+  1.8.4   K Hoang      11/01/2022 Fix libb64 compile error for ESP8266
  *************************************************************************************************************************************/
 
 #pragma once
 
-#define ETHERNET_WEBSERVER_VERSION          "EthernetWebServer v1.8.3"
+#define ETHERNET_WEBSERVER_VERSION          "EthernetWebServer v1.8.4"
 
 #define ETHERNET_WEBSERVER_VERSION_MAJOR    1
 #define ETHERNET_WEBSERVER_VERSION_MINOR    8
-#define ETHERNET_WEBSERVER_VERSION_PATCH    3
+#define ETHERNET_WEBSERVER_VERSION_PATCH    4
 
-#define ETHERNET_WEBSERVER_VERSION_INT      1008003
+#define ETHERNET_WEBSERVER_VERSION_INT      1008004
 
 #define USE_NEW_WEBSERVER_VERSION       true