+++ src/libb64/cencode.c
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.8.3
+  Version: 1.8.4
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -24,7 +24,10 @@
   1.8.1   K Hoang      24/12/2021 Fix bug
   1.8.2   K Hoang      27/12/2021 Fix wrong http status header bug
   1.8.3   K Hoang      28/12/2021 Fix authenticate issue caused by libb64
+  1.8.4   K Hoang      11/01/2022 Fix libb64 compile error for ESP8266
  *****************************************************************************************************************************/
+
+#if !(ESP32 || ESP8266)
  
 #include "cencode.h"
 
@@ -147,3 +150,5 @@ int base64_encode_chars(const char* plaintext_in, int length_in, char* code_out)
 
   return len + base64_encode_blockend((code_out + len), &_state);
 }
+
+#endif